requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
antiorm==1.2.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
db==0.1.1
db-sqlite3==0.0.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pyodbc==5.2.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
yarl==1.18.3
